Compatible with jenkins 3.19 and pipelines (#3)
* re-enabled the mailer.
fix the code to allow this test
`testNoTestsHaveQuarantineActionForStandardPublisher` and its functionality to work.

* Fix 6 code analysis issues.
Building this project no longer needs the
`findbungs.skip=true` flag  just use `mvn clean install`

* - update the readme.md

* - some changes for when the parent job is a Workflow job - this is for pipelines
