+++ readme.md
@@ -2,18 +2,13 @@
 MODIFICATIONS IN THE BRANCH
 ===========================
 
-Purpose =
+Purpose of this forking:
   - make it compatible with the latest version oj Jenkins 2.138.3 at the time of writing this
-  - make it compatible with rg.jenkins-ci.plugins 3.19
+  - make it compatible with org.jenkins-ci.plugins 3.19
   - make it compatible with pipelines
 
 Known issues:
-  - had to disable the mailer might get it to work if I look into it, disabled for now; had to disable the 3 tests that were using mailer
   - had to disable the maven test injection didn't care at this time about those tests might need re-write
-  - IMPORTANT: build using the `findbungs.skip=true` flag as the new code analysis found 6 "bugs" I don't have time to address now
-      e.g. mvn clean install -Dfindbugs.skip=true
-
-
 
 Jenkins Quarantine Plugin
 =========================
